requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.39.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
